Threshold HCR with Fmax scaled by Age Structure

Compute an allowable fishing mortality rate from a threshold HCR parameterized by f_min, f_max, lrp, and urp, but scales f_max by an age structure summary statistic relative to the value of that summmary statistic for some reference age structure.

as_scalar_threshold_f(x, naa, ref_naa, as_func, ..., f_min, f_max, lrp, urp)

Arguments

x

input to HCR (e.g. SSB or, SSB/B40)

naa

an age structure (e.g. numbers at age or biomass at age)

ref_naa

am age structure to compare `naa` to

as_func

a function that compute an age structure statistic (must take naa as its first argument)

...

additional arguments to pass to `as_func(naa, ...)`

f_min

F to use when x < lrp

f_max

F to use when x > urp

lrp

value of x below which f_min applies

urp

value of x above which f_max applies