Beverton-Holt stock recruitment relationship #' Defines a function factory for yieding recruitment from a Bevrton-Holt stock recruitment relationship, parameterized using steepeness. Initial call to this function sets up a new function that takes the current SSB as input, and returns the predicted recruitment from the BH SRR defined by h, R0, and S0.

Beverton-Holt stock recruitment relationship #' Defines a function factory for yieding recruitment from a Bevrton-Holt stock recruitment relationship, parameterized using steepeness. Initial call to this function sets up a new function that takes the current SSB as input, and returns the predicted recruitment from the BH SRR defined by h, R0, and S0.

Usage

beverton_holt(h, R0, S0, sigR, seed)

Arguments

h

steepness (0.2 <= h <= 1.0)

R0

unfished recruitment

S0

unfished spawning biomass

Examples

if (FALSE) { # \dontrun{
 bevholt <- beverton_holt(0.7, 25, 300)
 bevholt(ssb=185) # 23.43891
} # }