Compute Spawning Potential Ratio (SPR)

Compute SPR, the ratio of spawning biomass under a given leve of fishing mortality relative to unfished spawning biomass.

Usage

compute_spr(nages, mort, mat, waa, sel, ret, F)

Arguments

nages

number of ages in age structure

mort

instantaneous natural mortality rate

mat

maturity-at-age vector

waa

weight-at-age vector

sel

total selectivity-at-age vector

ret

total retention-at-age vector

F

instantenous fishing mortality rate