Get Reference Points from MSE Simulations

Derive fishing mortality and biomass reference points from completed MSE simulations.

Usage

get_reference_points(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  seed_list
)

Arguments

model_runs

list of completed MSE simulations

extra_columns

additional columns that should be appended to the resultant data frame

hcr_filter

vector of HCR names to process (must match names in `extra_columns`)

om_filter

vector of OM names to process (must match names in `extra_columns`)

seed_list

simulation seeds used in `model_runs`

Details

Note that this function hasn't been tested when multiple MSE simulations are present in the `model_runs` list.