Compute all Performance Metrics

Wrapper function to compute all performance metrics (identified below). - average annual catch - total catch - proportion of years catch exceeds a threshold - average ssb - proportion of years SSB is below a threshold - average age of the population - average ABI of the population (Griffiths et al. 2023) - average annual catch variation - average proportion of catch that is "large" - average proportion of population that is "old" - annual relative value - annual dynamic value - number of years required for population to delcine below 0.2B0 - number of years required for population to exceed below 0.4B0

Usage

performance_metric_summary(
  model_runs,
  extra_columns,
  dem_params,
  ref_naa,
  hcr_filter,
  om_filter,
  interval_widths,
  time_horizon = c(65, NA),
  extra_filter = NULL,
  relative = NULL,
  summarise_by = c("om", "hcr"),
  summary_out = TRUE,
  metric_list = "all"
)

Arguments

model_runs

list of completed MSE simulations runs

extra_columns

data.frame specifying names for OM and HCR to attach to each model_run (see `bind_mse_outputs` for more details)

dem_params

demographic parameters matrices from OM

ref_naa

reference age structure for ABI computation

hcr_filter

vector of HCR names to calculate metric over

om_filter

vector of OM names to calculate metric over

interval_widths

confidence intevrals to compute

extra_filter

an additional set of filters to apply before computing medians and confidence intervals

relative

a management procedure to compute metric relative to

summarise_by

vector of columns to summarise metric by

summary_out

whether to output data summarised by `ggdist` or full data

metric_list

vector of names of performance metrics to compute - `avg_catch` -> average annual catch - `tot_catch` -> total catch - `prop_years_high_catch` -> proportion of years catch exceeds a threshold - `avg_ssb` -> average ssb - `prop_years_lowssb` -> proportion of years SSB is below a threshold - `avg_age` -> average age of the population - `avg_abi` -> average ABI of the population (Griffiths et al. 2023) - `avg_variation` -> average annual catch variation - `avg_catch_lg` -> average proportion of catch that is "large" - `avg_pop_old` -> average proportion of population that is "old" - `annual_value` -> annual relative value - `dynamic_value` -> annual dynamic value - `crash_time` -> number of years required for population to delcine below 0.2B0 - `recovery_time` -> number of years required for population to exceed below 0.4B0 - `all` -> compute all of the above metrics

Value

list of dataframes. One element for each individual metric, and one element with all metrics