Run Management Strategy Evaluation

Run an MSE simulation loop.

Usage

run_mse(om, mp, mse_options, nyears_input = NA, seed = 1120, file_suffix = "")

Arguments

om

an operating model list object (should minimally contain a list object with demographic parameters, an array of initial NAAs, a list object specifying recruitment, and a list of model_options).

mp

a list object specifying the management procedure to be applied during the projection period

nyears_input

number of years to simulate forward (will override the dimensions of the demagraphic parameters matrices defined in the OM)

seed

random seed

...

parameters to pass to the `hcr` function

run_estimation

whether to run the estimation procedure or not. If FALSE, management procedures will use OM outputs rather than EM estimates to compute future harvest from the management procedure function.

spinup_years

number of years before estimation process should begin