Simulate Total Allowable Catch Projections

Determine the total allowable catch (TAC) in the next simulation year based on the fishing mortality rate identified by a harvest control rule (HCR). TAC is computed by projecting the population forward one year under the specific level of F and computing the the amount of catch that would be taken from that projected population structure based on the joint fishery selectivity curve.

Usage

simulate_TAC(
  hcr_F,
  naa,
  recruitment,
  joint_sel,
  dem_params,
  hist_abc,
  hcr_options,
  options
)

Arguments

hcr_F

fishing mortality rate identified by HCR

naa

numbers-at-age vector (dimensions [1, nages, nsexes, nregions])

recruitment

projected recruitment in next year

dem_params

demographic parameter matrices subsetted to 1 year

hcr_options

list of additional management options such as stability constraints and harvest caps

options

list of additional ABC/TAC simulation options such as ABC-TAC reduction levels and attainment functions.

join_sel

joint fishery selectivity

hist_tac

TAC from the previous year