Simulate OM Observation Data for TMB Assessment Model

Format data and observations from an `afscOM` operating model and update the model data and parameters required by the `SpatialSablefishAssessment` TMB assessment model.

Usage

simulate_em_data(
  nyears,
  dem_params,
  land_caa,
  survey_indices,
  fxfish_caa_obs,
  twfish_caa_obs,
  ll_ac_obs,
  tw_ac_obs,
  ll_srv_indic,
  tw_srv_indic,
  model_options,
  added_years = 1,
  file_suffix = ""
)

Arguments

nyears

the number of years of data that will be passed to the model

dem_params

demographic parameter matrices subsetted to 1 year

land_caa

nyears worth of landed catch-at-age data (dimensins [1, nages, nsexes, nregions, nfleets])

survey_indices

nyears worth of survey indices (LL RPN, LL RPW, and TW RPW)

fxfish_caa_obs

nyears worth of catch-at-age observation from the fixed gear fishery

twfish_caa_obs

nyears worth of catch-at-age observation from the trawl gear fishery

ll_ac_obs

nyears worth of age composition observations frmo the longline survey

tw_ac_obs

nyears worth of age composition observations frmo the trawl survey

model_options

list of model options provided to the OM

file_suffix

suffix to append to saved outputs

added_year

the number of new years of data being added (should usually be 1)

Details

This function requires the existence of 'data/sablefish_em_data_2022.RDS', and 'data/sablefish_em_par_2022.RDS' files, which come packaged with the `SablefishMSE` codebase.